# api/openapi.yaml
openapi: 3.0.3
info:
  title: Runcriticon API
  version: 0.1.0
  description: API REST de Runcriticon
servers:
  - url: /api
    description: Mismo origen que la SPA (cookie de sesión first-party, sin CORS)
...
security:
  - cookieAuth: []
paths:
...
  /grupos/{grupoId}/overrides/{alumnoId}:
    put:
      summary: Ajustar a mano la pertenencia de un alumno a un grupo
      description: |
        Mete o saca a un alumno de un grupo **sin tocar sus tags**, para gestionar excepciones sin
        ensuciar la taxonomía. El ADMIN o el ENTRENADOR.

        `incluido: true` lo mete aunque no cumpla el filtro; `incluido: false` lo saca aunque lo
        cumpla. La excepción **prevalece sobre cambios posteriores de los tags del alumno**: no se
        recalcula ni se pierde, solo se quita a mano con el DELETE.

        Idempotente: repetir la misma llamada deja el mismo estado, y llamarla con el valor contrario
        voltea la excepción sin necesidad de borrarla antes. Devuelve el detalle del grupo ya
        recalculado, para no obligar a una segunda consulta.

        404 si el grupo no existe o es de otro club, y también si el alumno no existe, es de otro
        club o no es alumno.
      operationId: ajustarPertenenciaAGrupo
      tags: [grupos]
      parameters:
        - $ref: '#/components/parameters/GrupoIdPath'
        - $ref: '#/components/parameters/AlumnoIdEnGrupo'
      requestBody:
        $ref: '#/components/requestBodies/GroupOverrideBody'
      responses:
        '200':
          $ref: '#/components/responses/GrupoDetalle'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '500':
          $ref: '#/components/responses/ErrorServerError'
    delete:
      summary: Quitar el ajuste manual y devolver al alumno al filtro
      description: |
        Borra la excepción: a partir de ahí, que el alumno pertenezca o no al grupo lo decide otra vez
        el filtro de tags. El ADMIN o el ENTRENADOR.

        Idempotente: devuelve 204 tanto si había excepción como si no. No comprueba el alumno —una
        excepción de alguien que ya no existe o que dejó de ser alumno tiene que poder limpiarse—, así
        que el 404 solo aparece si el grupo no existe o es de otro club.
      operationId: quitarAjusteDePertenencia
      tags: [grupos]
      parameters:
        - $ref: '#/components/parameters/GrupoIdPath'
        - $ref: '#/components/parameters/AlumnoIdEnGrupo'
      responses:
        '204':
          description: Ajuste manual quitado (o no había ninguno)
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '500':
          $ref: '#/components/responses/ErrorServerError'
...
  /planes:
    get:
      summary: Listar los planes en borrador de un grupo
      description: |
        Devuelve los planes semanales en estado BORRADOR de un grupo, ordenados por semana. El
        ENTRENADOR con relación con el grupo. No trae sesiones ni personalizaciones —solo id, semana
        y estado—: es el listado que alimenta "mis planes en borrador por grupo", no el detalle.

        `grupoId` inexistente o ajeno al entrenador devuelve una lista vacía, no 404: este módulo solo
        conoce su propia proyección de pertenencia, no si el grupo existe de verdad en `club_taxonomia`.
      operationId: listarPlanesEnBorrador
      tags: [planes]
      parameters:
        - name: grupoId
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/GroupId'
      responses:
        '200':
          $ref: '#/components/responses/PlanesListados'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '500':
          $ref: '#/components/responses/ErrorServerError'
    post:
      summary: Crear un plan semanal en borrador
      description: |
        Crea un plan en borrador para un grupo. Solo el ENTRENADOR, y solo si tiene relación con ese
        grupo (asignación de entrenadores a grupos con autorización de publicación) — un entrenador
        sin relación con el grupo recibe 403, igual que
        si el grupo no existiera: este módulo no distingue ambos casos.

        `semana` debe ser el lunes de la semana del plan; cualquier otro día devuelve 400.
      operationId: crearPlanEnBorrador
      tags: [planes]
      requestBody:
        $ref: '#/components/requestBodies/CreatePlanBody'
      responses:
        '201':
          $ref: '#/components/responses/PlanCreado'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '500':
          $ref: '#/components/responses/ErrorServerError'
  /planes/{planId}:
    get:
      summary: Consultar el detalle de un plan, con sus sesiones
      description: |
        Devuelve el plan completo con sus sesiones (editor de sesión) — a diferencia de
        `GET /planes`, que solo trae el resumen del listado. Solo el ENTRENADOR con relación con el
        grupo del plan; sin relación (o plan inexistente / de otro club) da 403, no 404: este módulo
        no distingue ambos casos, igual que en `POST /planes`.
      operationId: consultarPlan
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
      responses:
        '200':
          $ref: '#/components/responses/PlanDetalle'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '500':
          $ref: '#/components/responses/ErrorServerError'
  /planes/{planId}/sesiones:
    post:
      summary: Añadir una sesión al plan
      description: |
        Crea una sesión para un día concreto de la semana del plan (editor de sesión de
        entrenamiento: tipo, volumen —distancia o tiempo, nunca los dos—, ritmo y notas). `tipo` sale
        del catálogo cerrado del glosario; `DESCANSO` no admite `volumen` ni `ritmo`.

        En el MVP el ritmo solo se introduce como `ABSOLUTO` — el modelo ya admite `RELATIVO` porque
        el dominio lo soporta desde que se creó el plan semanal en borrador, pero es la UI de ritmos
        relativos a marcas quien lo activará.

        `dia` debe caer dentro de la semana del plan (400) y no puede coincidir con el de otra sesión
        ya creada (409, una sesión por día). Solo el ENTRENADOR con relación vigente con el grupo del
        plan (403 si no, igual que al crearlo).
      operationId: anadirSesion
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
      requestBody:
        $ref: '#/components/requestBodies/SessionBody'
      responses:
        '201':
          $ref: '#/components/responses/SesionCreada'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '500':
          $ref: '#/components/responses/ErrorServerError'
  /planes/{planId}/sesiones/{sesionId}:
    put:
      summary: Editar tipo, volumen, ritmo y notas de una sesión
      description: |
        Sustituye tipo, volumen, ritmo y notas de una sesión existente. **Sin `dia`**: el editor no
        permite mover una sesión de día — para eso se borra y se crea otra. 404 si la sesión no
        existe en el plan.
      operationId: editarSesion
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/SesionIdPath'
      requestBody:
        $ref: '#/components/requestBodies/SessionUpdateBody'
      responses:
        '200':
          $ref: '#/components/responses/SesionActualizada'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '500':
          $ref: '#/components/responses/ErrorServerError'
    delete:
      summary: Eliminar una sesión del plan
      description: Elimina la sesión. 404 si no existe en el plan.
      operationId: eliminarSesion
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/SesionIdPath'
      responses:
        '204':
          description: Sesión eliminada
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '500':
          $ref: '#/components/responses/ErrorServerError'
  /planes/{planId}/publicacion:
    post:
      summary: Publicar el plan semanal a su grupo
      description: |
        Publica el plan: pasa a `PUBLICADO`, congela el snapshot de alumnos resueltos en ese
        momento (ADR-0002 D5, `plan_snapshot_alumno`) y emite `PlanPublicado`. A partir de aquí el
        plan deja de aceptar sesiones nuevas ni ediciones — 409 `PLAN_ALREADY_PUBLISHED` en ese caso,
        igual que al intentar publicarlo dos veces.

        Un plan sin ninguna sesión no se puede publicar (409 `PLAN_WITHOUT_SESSIONS`): publicar una
        semana en blanco es siempre un error del entrenador. Solo el ENTRENADOR con relación vigente
        con el grupo del plan (403, mismo criterio que el resto de operaciones sobre el plan).

        Si la proyección de membresía del grupo está atrasada más de lo tolerable (ADR-0009 D9,
        fail-closed a 60 s), devuelve 503 `PROJECTION_STALE`: publicar ahora arriesgaría un snapshot
        incompleto. Es un estado transitorio — reintentar en unos segundos suele resolverlo.
      operationId: publicarPlan
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
      responses:
        '200':
          $ref: '#/components/responses/PlanPublicado'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '500':
          $ref: '#/components/responses/ErrorServerError'
        '503':
          $ref: '#/components/responses/ErrorServiceUnavailable'
  /planes/{planId}/sesiones/{sesionId}/personalizaciones/{alumnoId}:
    put:
      summary: Aplicar o sustituir la personalización de un alumno en una sesión
      description: |
        Crea o sustituye el override de esta sesión para este alumno: reemplazo completo de
        tipo/volumen/ritmo/notas, nunca combina campo a campo con la sesión base. Única por
        (plan, sesión, alumno) — repetir la llamada sustituye el override anterior, no lo acumula.

        Permitido tanto con el plan en `BORRADOR` como ya `PUBLICADO` (a diferencia de las
        operaciones sobre sesiones, que la publicación congela): personalizar un plan ya publicado es
        el caso de uso principal. `alumnoId` debe pertenecer al grupo del plan si está en `BORRADOR`,
        o estar en su snapshot congelado si ya está `PUBLICADO` (409 `STUDENT_NOT_IN_PLAN` si no).

        Devuelve el plan completo recalculado, para no obligar a una segunda consulta (mismo criterio
        que `ajustarPertenenciaAGrupo`). Solo el ENTRENADOR con relación vigente con el grupo del plan.

        404 si la sesión no existe en el plan (`SESSION_NOT_FOUND`).
      operationId: aplicarPersonalizacion
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/SesionIdPath'
        - $ref: '#/components/parameters/AlumnoIdPersonalizado'
      requestBody:
        $ref: '#/components/requestBodies/PersonalizationBody'
      responses:
        '200':
          $ref: '#/components/responses/PlanDetalle'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorConflict'
        '500':
          $ref: '#/components/responses/ErrorServerError'
    delete:
      summary: Retirar la personalización de un alumno en una sesión
      description: |
        Retira el override, si existía. A diferencia de `quitarAjusteDePertenencia` (que es 204
        siempre, aunque no hubiera nada), aquí se distingue: 404 `PERSONALIZATION_NOT_FOUND` si el
        alumno no tenía personalización en esa sesión. Solo el ENTRENADOR con relación vigente con el
        grupo del plan.
      operationId: retirarPersonalizacion
      tags: [planes]
      parameters:
        - $ref: '#/components/parameters/PlanIdPath'
        - $ref: '#/components/parameters/SesionIdPath'
        - $ref: '#/components/parameters/AlumnoIdPersonalizado'
      responses:
        '204':
          description: Personalización retirada
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '500':
          $ref: '#/components/responses/ErrorServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
      description: |
        Sesión autenticada por cookie httpOnly first-party (Spring Session). El navegador la adjunta
        automáticamente; el cliente nunca la lee ni la envía a mano. Los endpoints de entrada anónima
        (login, activación, magic link, reseteo) no la requieren (security: []).
...
  parameters:
...
    GrupoIdPath:
      name: grupoId
      in: path
      required: true
      description: ID del grupo del club
      schema:
        $ref: '#/components/schemas/GroupId'
    AlumnoIdEnGrupo:
      name: alumnoId
      in: path
      required: true
      description: ID del alumno sobre el que se ajusta la pertenencia al grupo
      schema:
        type: string
        format: uuid
...
    PlanIdPath:
      name: planId
      in: path
      required: true
      description: ID del plan semanal
      schema:
        $ref: '#/components/schemas/PlanId'
    SesionIdPath:
      name: sesionId
      in: path
      required: true
      description: ID de la sesión dentro del plan
      schema:
        $ref: '#/components/schemas/TrainingSessionId'
    AlumnoIdPersonalizado:
      name: alumnoId
      in: path
      required: true
      description: ID del alumno del grupo sobre el que se aplica o retira la personalización
      schema:
        type: string
        format: uuid
...
  requestBodies:
...
    GroupOverrideBody:
      description: Sentido del ajuste manual — dentro del grupo o fuera de él
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GroupOverrideRequest'
    CreatePlanBody:
      description: Grupo y semana del plan en borrador
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatePlanRequest'
    SessionBody:
      description: Día, tipo, volumen, ritmo y notas de la sesión a crear
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TrainingSessionRequest'
    SessionUpdateBody:
      description: Tipo, volumen, ritmo y notas de la sesión — sin día, que no se puede mover
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TrainingSessionUpdateRequest'
    PersonalizationBody:
      description: Override completo de la sesión para el alumno, más el mensaje opcional
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PersonalizationRequest'
...
  responses:
...
    ErrorBadRequest:
      description: Entrada inválida (campo requerido vacío, formato incorrecto)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INVALID_INPUT
            field: email
            message: El email no tiene un formato válido
    ErrorUnauthorized:
      description: No autenticado; la sesión ha expirado o no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: No autenticado
    ErrorForbidden:
      description: El principal no tiene permiso (cuerpo neutro sin detalles)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Acceso denegado
    ErrorNotFound:
      description: Recurso no encontrado o no pertenece al club del principal
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Recurso no encontrado
    ErrorConflict:
      description: La operación choca con el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Ya existe un usuario con ese email en el club
...
    ErrorServerError:
      description: |
        Error interno del servidor (red de seguridad de GlobalRestExceptionHandler). Cuerpo
        ErrorResponse con `code=INTERNAL_ERROR` y mensaje neutro: el detalle del fallo va al log,
        nunca al body. Puede devolverlo cualquier endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Ha ocurrido un error inesperado
    ErrorServiceUnavailable:
      description: |
        Indisponibilidad temporal, distinta de un error del cliente: la operación es correcta pero no
        se puede completar ahora con garantías (ej. `PROJECTION_STALE`, ADR-0009 D9). Reintentar en
        unos segundos suele resolverlo.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: PROJECTION_STALE
            message: La membresía del grupo está desactualizada; inténtalo de nuevo en unos segundos
...
    GrupoDetalle:
      description: Grupo con sus miembros —y por qué lo son— y sus exclusiones manuales
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GroupDetailResponse'
...
    PlanCreado:
      description: Plan semanal creado en estado BORRADOR
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PlanResponse'
    PlanesListados:
      description: Planes en borrador del grupo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PlanesResponse'
    PlanDetalle:
      description: Plan semanal completo, con sus sesiones
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PlanDetalleResponse'
    SesionCreada:
      description: Sesión creada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TrainingSessionResponse'
    SesionActualizada:
      description: Sesión después de la edición
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TrainingSessionResponse'
    PlanPublicado:
      description: El plan tras publicarse, con el tamaño del snapshot congelado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PublicacionResponse'
...
  schemas:
...
    GroupId:
      type: string
      format: uuid
      description: ID de un grupo del club
      example: 018f9a2b-1c3d-7e4f-8a9b-0c1d2e3f4a5d
...
    PlanId:
      type: string
      format: uuid
      description: ID de un plan semanal
      example: 018f9a2b-1c3d-7e4f-8a9b-0c1d2e3f4a5e
    CreatePlanRequest:
      type: object
      description: |
        Alta de un plan en borrador para un grupo. `semana` debe ser el lunes de esa semana — cualquier
        otro día se rechaza con 400.
      required: [grupoId, semana]
      properties:
        grupoId:
          $ref: '#/components/schemas/GroupId'
        semana:
          type: string
          format: date
          description: Lunes de la semana del plan
          example: '2026-08-17'
    PlanResponse:
      type: object
      description: |
        Un plan semanal, en borrador o publicado. Sin sesiones ni personalizaciones: es el resumen que
        alimenta el listado (`GET /planes`) — el detalle completo lo trae `GET /planes/{planId}`
        (`PlanDetalleResponse`).
      required: [id, grupoId, semana, estado]
      properties:
        id:
          $ref: '#/components/schemas/PlanId'
        grupoId:
          $ref: '#/components/schemas/GroupId'
        semana:
          type: string
          format: date
          example: '2026-08-17'
        estado:
          type: string
          enum: [BORRADOR, PUBLICADO]
          example: BORRADOR
    PlanesResponse:
      type: object
      description: Planes en borrador de un grupo, ordenados por semana.
      required: [planes]
      properties:
        planes:
          type: array
          items:
            $ref: '#/components/schemas/PlanResponse'
    TrainingSessionId:
      type: string
      format: uuid
      description: ID de una sesión dentro de un plan
      example: 018f9a2b-1c3d-7e4f-8a9b-0c1d2e3f4a5f
...
    TrainingSessionResponse:
      type: object
      required: [id, dia, tipo]
      properties:
        id:
          $ref: '#/components/schemas/TrainingSessionId'
        dia:
          type: string
          format: date
        tipo:
          type: string
          enum: [RODAJE, SERIES, TEMPO, TIRADA_LARGA, FARTLEK, CUESTAS, PROGRESIVO, FUERZA_CROSS, COMPETICION, DESCANSO]
        volumen:
          $ref: '#/components/schemas/Volumen'
        ritmo:
          $ref: '#/components/schemas/Ritmo'
        notas:
          type: string
...
    PlanDetalleResponse:
      type: object
      description: |
        El plan completo con sus sesiones y sus personalizaciones vigentes. Se
        devuelve recalculado tras `PUT`/`DELETE` de una personalización, para no obligar a una
        segunda consulta (mismo criterio que `GrupoDetalle` en `ajustarPertenenciaAGrupo`).
      required: [id, grupoId, semana, estado, sesiones]
      properties:
        id:
          $ref: '#/components/schemas/PlanId'
        grupoId:
          $ref: '#/components/schemas/GroupId'
        semana:
          type: string
          format: date
          example: '2026-08-17'
        estado:
          type: string
          enum: [BORRADOR, PUBLICADO]
          example: BORRADOR
        sesiones:
          type: array
          items:
            $ref: '#/components/schemas/TrainingSessionResponse'
        personalizaciones:
          type: array
          description: Overrides vigentes por alumno, permitidos tanto en BORRADOR como en PUBLICADO.
          items:
            $ref: '#/components/schemas/PersonalizationResponse'
...
    GroupOverrideRequest:
      type: object
      description: |
        Sentido del ajuste manual: `true` mete al alumno en el grupo aunque no cumpla el filtro,
        `false` lo saca aunque lo cumpla. No hay tercer estado — para volver a dejarlo en manos del
        filtro se borra la excepción con el DELETE.
      required: [incluido]
      properties:
        incluido:
          type: boolean
          example: true
...
    GroupDetailResponse:
      type: object
      description: |
        Un grupo con su composición actual. `total` es el tamaño de `miembros`; las exclusiones no
        cuentan, así que el desglose que se pinta en pantalla es «los del filtro, menos los excluidos,
        más los incluidos a mano».

        No se pagina ni se trunca: el dimensionado previsto es de un par de cientos de alumnos por
        club. `miembros` viene ordenado por nombre.
      required: [id, nombre, valores, total, miembros, excluidos]
      properties:
        id:
          $ref: '#/components/schemas/GroupId'
        nombre:
          $ref: '#/components/schemas/GroupName'
        valores:
          type: array
          items:
            $ref: '#/components/schemas/TagValueId'
        total:
          type: integer
          format: int32
          minimum: 0
          example: 13
        miembros:
          type: array
          items:
            $ref: '#/components/schemas/GroupMembershipResponse'
        excluidos:
          type: array
          items:
            $ref: '#/components/schemas/GroupExclusionResponse'
...
    ErrorResponse:
      type: object
      description: |
        Error estructurado para respuestas 4xx y 5xx.
        El frontend traduce `code` a mensaje localizado; `field` solo aparece en errores de entrada.
      required: [code, message]
      properties:
        code:
          type: string
          description: Código estable para que el frontend lo traduzca
          example: INVALID_INPUT
        field:
          type: string
          nullable: true
          description: |
            Campo que originó el error, cuando hay uno concreto. Habitual en los 400 de validación,
            pero no exclusivo: `DUPLICATE_LABEL` lo lleva en un 409 para señalar qué campo choca.
          example: email
        message:
          type: string
          description: Mensaje legible (orientativo, no usar para lógica de cliente)
          example: El email no tiene un formato válido